// frontend/src/client/types.gen.ts
export type ReviewPublic = {
    /**
     * Id
     */
    id: string;
    /**
     * Rating
     */
    rating: number;
    /**
     * Comment
     */
    comment: string | null;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string | null;
    /**
     * User Id
     */
    user_id: string;
    /**
     * Movie Id
     */
    movie_id: string;
};

/**
 * ReviewUpdate
 */
export type ReviewUpdate = {
    /**
     * Rating
     */
    rating?: number | null;
    /**
     * Comment
     */
    comment?: string | null;
};